apps/backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn>=0.14.0,<0.29.0
starlette==0.37.2
python-multipart==0.0.9
email-validator==2.2.0
slowapi>=0.1.8

# Database
sqlalchemy==2.0.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT[crypto]>=2.8.0
passlib==1.7.4
bcrypt==4.3.0
cryptography==45.0.3

# Data Validation
pydantic>=2.7.4,<3.0
pydantic-core>=2.18.4,<3.0

# Observability
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-sqlalchemy==0.60b1
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-semantic-conventions==0.60b1
prometheus-client==0.20.0
structlog==25.4.0

# Background Jobs & Scheduling
apscheduler==3.10.4

# AI & ML — LangChain multi-provider
langchain-core>=0.3.0,<1.0
langchain-openai>=0.3.0,<1.0
langchain-anthropic>=0.3.0,<1.0
langchain-google-genai>=2.1.0,<5.0
openai>=1.25.0
numpy>=1.24.4
scikit-learn>=1.5.0
scipy>=1.12.0

# AWS
boto3>=1.34.0
botocore>=1.34.0

# Utilities
python-dotenv==1.0.1
requests==2.32.3
python-dateutil==2.9.0.post0

# Redis for metrics persistence
redis>=4.5.0

# ML Explainability
shap>=0.49.0

# ONNX Runtime (lightweight ML inference)
onnxruntime>=1.17.0

# Async HTTP client (outbound webhooks, pull ingestion)
httpx>=0.27.0

# MCP Server (Model Context Protocol)
mcp>=1.2.0

# Kafka (optional, for event ingestion)
confluent-kafka>=2.0.0